Move SearchService to new API endpoint
